--- src/main/kotlin/blackjack/domain/CardResult.kt
@@ -1,7 +1,0 @@
-package blackjack.domain
-
-data class CardResult(
-    val participant: Participant,
-    val cards: List<Card>,
-    val scoreSum: Int,
-)
