+++ src/main/kotlin/blackjack/domain/card/CardNumber.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class CardNumber(val score: Int) {
     ACE(1),
