+++ src/main/kotlin/blackjack/domain/participant/Player.kt
@@ -1,4 +1,6 @@
-package blackjack.domain
+package blackjack.domain.participant
+
+import blackjack.domain.card.Card
 
 class Player(name: String, private val needToDraw: () -> Boolean = { true }) : Participant(name) {
     override fun getFirstOpenCards(): List<Card> = getCards()
