+++ src/main/kotlin/blackjack/domain/result/BlackjackResult.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.result
 
 data class BlackjackResult(
     val cardResults: List<CardResult>,
