+++ src/main/kotlin/blackjack/domain/result/GameResult.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.result
 
 enum class GameResult {
     WIN, DRAW, LOSE;
