+++ src/main/kotlin/blackjack/domain/result/MatchResult.kt
@@ -1,4 +1,6 @@
-package blackjack.domain
+package blackjack.domain.result
+
+import blackjack.domain.participant.Participant
 
 data class MatchResult(
     val participant: Participant,
