+++ src/test/kotlin/blackjack/domain/CardNumberTest.kt
@@ -1,5 +1,6 @@
 package blackjack.domain
 
+import blackjack.domain.card.CardNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
